dom/nodes/ProcessingInstruction-in-doctype.xhtml

<!DOCTYPE html [<?x y?>]><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>XML: Processing instruction in doctype internal subset</title>
    <link rel="help" href="https://www.w3.org/TR/2004/REC-xml-infoset-20040204/#infoitem.pi"/>
    <!-- "[parent] The document, element, or document type declaration information item
    which contains this information item in its [children] property. " -->
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_equals(document.documentElement.previousSibling, document.firstChild);
        assert_equals(document.firstChild.nodeType, Node.DOCUMENT_TYPE_NODE);
      });
    </script>
  </body>
</html>
