third_party/blink/web_tests/external/wpt/dom/nodes/ProcessingInstruction-in-doctype-expected.txt

This is a testharness.js-based test.
[FAIL] XML: Processing instruction in doctype internal subset
  assert_equals: expected DocumentType node but got ProcessingInstruction node with target "x" and data "y"
Harness: the test ran to completion.
